% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrank.R
\name{blockrank.diablo}
\alias{blockrank.diablo}
\title{Run BlockRank for multi-omic model}
\usage{
blockrank.diablo(diablo.model)
}
\arguments{
\item{diablo.model}{Trained multi-omics mixomics DIABLO model}
}
\value{
A list of numeric vectors containing BlockRank scores.
One list item per data block
}
\description{
Takes a DIABLO model trained on multi-omic data using the mixomics package
and calculates the BlockRank scores for each feature.
This function can also calculate scores for block.plsda models.
}
\examples{
\dontrun{
diablo.blockrank.scores <- blockrank.diablo(diablo.model)
}
}
