% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrank.R
\name{blockrank.splsda}
\alias{blockrank.splsda}
\title{Run BlockRank for single-omic model}
\usage{
blockrank.splsda(splsda.model)
}
\arguments{
\item{splsda.model}{trained single-omic mixomics sPLS-DA model}
}
\value{
A list of 1 numeric vector containing BlockRank scores. One list item for single data block
}
\description{
Takes an sPLS-DA model trained on single-omic data using the mixomics package
and calculates the BlockRank scores for each feature.
This function can also calculate scores for plsda models.
}
\examples{
\dontrun{
splsda.blockrank.scores <- blockrank.splsda(splsda.model)
}
}
