% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_mapping.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Convert a list of mouse genes to their human orthologues.}
\usage{
convertMouseGeneList(mouse.genes)
}
\arguments{
\item{mouse.genes}{List of mouse gene symbols.}
}
\value{
Human orthologue gene symbols (not necessarily ordered).
}
\description{
Convert a list of mouse genes to their human orthologues.
}
