% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{diablo.selection.stability}
\alias{diablo.selection.stability}
\title{Extract DIABLO feature selection stability}
\usage{
diablo.selection.stability(diablo.perf, comp, block)
}
\arguments{
\item{diablo.perf}{Performance evaluation of the DIABLO model generated by
mixOmics perf.}

\item{comp}{Component number for which to retrieve feature selection
stabilities.}

\item{block}{Block number for which to retrieve feature selection
stabilities.}
}
\value{
A data frame containing the  the features selected for the specificed
component and their relative stability, as a proportion of trained models
0-1.
}
\description{
Extract feature selection stabilities for a given component from a
performance validated mixOmics multi-omics DIABLO model. Feature stabilities
are an important indicator of the confidence that a selected feature is
predictive for the outcome classes in the model, and hence (in combination
with the loading weight) is likely to be of biological significance. Features
are retrieved for a single specified block and component.
}
\examples{
\dontrun{
diablo.selection.stability(perf.diablo.analysis, comp=1, block=1)
}
}
