% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrank.R
\name{.internal.blockrank.calculation}
\alias{.internal.blockrank.calculation}
\title{Internal BlockRank Calculation}
\usage{
.internal.blockrank.calculation(
  x.data,
  y.data,
  x.loadings,
  y.loadings,
  ncomp,
  nblock
)
}
\arguments{
\item{x.data}{List containing the centered and standardized original predictor matrix}

\item{y.data}{Matrix with the position of the outcome Y in the output list X}

\item{x.loadings}{List containing the estimated loadings for the variates.}

\item{y.loadings}{Matrix containing the estimated loadings for Y.}

\item{ncomp}{Number of components included in the model for each block}

\item{nblock}{Number of blocks of data}
}
\value{
A list of numeric vectors containing BlockRank scores. One list item per data block
}
\description{
Internal BlockRank Calculation
}
