% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{.plot.module.membership.correlation}
\alias{.plot.module.membership.correlation}
\title{Make a scatter plot of the degree of module membership vs the correlation
with an outcome variable.}
\usage{
.plot.module.membership.correlation(
  module,
  module.colors,
  gene.module.membership,
  gene.trait.significance,
  outcome.name = "outcome"
)
}
\arguments{
\item{module}{Module name to plot.}

\item{module.colors}{List of gene module colours.}

\item{gene.module.membership}{Previously calculated gene module membership.}

\item{gene.trait.significance}{Previously calculated gene trait significance.}

\item{outcome.name}{Name of the outcome being correlated.}
}
\value{
Scatter plot.
}
\description{
Make a scatter plot of the degree of module membership vs the correlation
with an outcome variable.
}
