% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMC.R
\name{.TSS.divide}
\alias{.TSS.divide}
\title{Divide}
\usage{
.TSS.divide(x)
}
\arguments{
\item{x}{Row to divide over.}
}
\value{
TSS scaled row.
}
\description{
Internal function for dividing over complete samples in Total Sum Scaling
(TSS).
}
