% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{export.matrix.as.network}
\alias{export.matrix.as.network}
\title{Export a covariance matrix, e.g. one generated using
`find.feature.associations` as data describing the association network and
suitable for e.g. loading into Cytoscape. The method can optionally take a
cutoff value, which is used to threshold (absolute) association scores. The
output is written to disk as a CSV file using the filename given.}
\usage{
export.matrix.as.network(
  m,
  filename,
  cutoff = NA,
  block.association = NULL,
  block.feature.count = 20
)
}
\arguments{
\item{m}{Symmetrical matrix of association scores.}

\item{filename}{Filename to which to write network output.}

\item{cutoff}{Optional value of which to filter associations.}

\item{block.association}{If the matrix does not contain equal number of
variables from each block, then a custom list of block associations can be
passed in.}

\item{block.feature.count}{The number of features selected from each block,
defaulting to 20.}
}
\value{
The long-form network data that is also written to disk
}
\description{
Export a covariance matrix, e.g. one generated using
`find.feature.associations` as data describing the association network and
suitable for e.g. loading into Cytoscape. The method can optionally take a
cutoff value, which is used to threshold (absolute) association scores. The
output is written to disk as a CSV file using the filename given.
}
\examples{
\dontrun{
export.matrix.as.network(mat, "network.csv")
export.matrix.as.network(find.feature.associations(diablo.trained.analysis, 3), filename="network.csv", cutoff=0.7)
}
}
