% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{get.diablo.top.loadings.1comp}
\alias{get.diablo.top.loadings.1comp}
\title{Get a dataframe indicating the top ranking loading factors for a trained mixOmics multi-omics DIABLO model for a
provided block from the model.}
\usage{
get.diablo.top.loadings.1comp(diablo.loadings, feature.count = 20)
}
\arguments{
\item{diablo.loadings}{DIABLO loadings for a single component model}

\item{feature.count}{The number of top loadings from each component to include.}
}
\value{
A data frame containing the top (by absolute magnitude) factor loadings for the single component, ordered by absolute weight.
}
\description{
Get a dataframe indicating the top ranking loading factors for a trained mixOmics multi-omics DIABLO model for a
provided block from the model.
}
\examples{
\dontrun{
get.diablo.top.loadings.1comp(diablo.tuned.analysis$loadings$metab)
}
}
