% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{get.loadings.table}
\alias{get.loadings.table}
\title{Get an sPLS-DA loadings table}
\usage{
get.loadings.table(splsda.analysis)
}
\arguments{
\item{splsda.analysis}{Trained mixOmics sPLS-DA model}
}
\value{
Data frame including ordered loadings over the first two components
of the model. Selected features are ordered by component one followed by
component two, and then from highest to lowest (in absolute terms).
}
\description{
Get a dataframe indicating the loading factors for a trained mixOmics sPLS-DA
model. Limited to two components.
}
\examples{
\dontrun{
get.loadings.table(splsda.analysis)
}
}
