% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMC.R
\name{low.count.removal}
\alias{low.count.removal}
\title{Prefilter omics analysis input data in count form (e.g. OTUs) to remove
features which have a total count less than a (small) proportion of the total
measured counts. The default threshold is one part in 10,000 (0.01\%) - this
is usually sufficient to remove very low-count variables, which will be
unreliable features for model prediction.}
\usage{
low.count.removal(otu.counts, percent = 0.01)
}
\arguments{
\item{otu.counts}{OTU count data frame of size n (sample) x p (OTU).}

\item{percent}{Cutoff chosen in percent, default to 0.01.}
}
\value{
Data frame of input data, filtered to omit features below the count
proportion threshold.
}
\description{
Prefilter omics analysis input data in count form (e.g. OTUs) to remove
features which have a total count less than a (small) proportion of the total
measured counts. The default threshold is one part in 10,000 (0.01\%) - this
is usually sufficient to remove very low-count variables, which will be
unreliable features for model prediction.
}
\examples{
\dontrun{
low.count.filter(raw.count)
}
}
