% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_mapping.R
\name{map.symbols.entrez.ids}
\alias{map.symbols.entrez.ids}
\title{Map gene symbols to Entrez IDs}
\usage{
map.symbols.entrez.ids(symbols, organism)
}
\arguments{
\item{symbols}{List of gene symbols.}

\item{organism}{The organism the symbols relate to. This can be any of the
supported common names, scientific names or shorthand shown when calling
\code{anRichmentMethods::organismLabels()}.}
}
\value{
Data frame containing pairs of symbols and Entrez IDs.
}
\description{
Map a list of symbols to their corresponding Entrez IDs. This wraps the
method from the anRichmentMethods library to convert a list of symbols to
entrez IDs.
}
\examples{
\dontrun{
gene.symbols.and.ids <- map.symbols.entrez.ids(gene.symbols, "mouse")
}
}
