% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{merge.feature.stability}
\alias{merge.feature.stability}
\title{Utility function to combine top loading factors extracted from a trained
multi-omics mixOmics (DIABLO) model with the feature stability, previously
retrieved from the performance evaluation using e.g.
diablo.selection.stability and create a nicely formatted data frame.}
\usage{
\method{merge}{feature.stability}(features, stability)
}
\arguments{
\item{features}{Data frame of top features, e.g. provided by
get.diablo.top.loadings().}

\item{stability}{Data frame of feature stability, e.g. provided by
diablo.selection.stability().}
}
\value{
Data frame combining top features with their measured stability.
}
\description{
Utility function to combine top loading factors extracted from a trained
multi-omics mixOmics (DIABLO) model with the feature stability, previously
retrieved from the performance evaluation using e.g.
diablo.selection.stability and create a nicely formatted data frame.
}
\examples{
\dontrun{
merge.feature.stability(get.diablo.top.loadings(diablo.tuned.analysis$loadings$metab), diablo.selection.stability(perf.diablo.analysis, comp=1, block=1))
}
}
