% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{module.membership.pie}
\alias{module.membership.pie}
\title{Create a pie chart of gene module membership}
\usage{
module.membership.pie(moduleColors)
}
\arguments{
\item{moduleColors}{Module colours per gene.}
}
\value{
ggplot2 pie chart.
}
\description{
Creates a well-presented pie chart of gene module membership, with the
modules represented by their actual colours.
}
