% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMC.R
\name{normalise.logit}
\alias{normalise.logit}
\title{Apply the standard logit normalisation to a data frame of omics input data.
This is intended to convert compositional data, e.g. proportional data in the
range 0..1, to Euclidean space which is most appropriate for the linear
models. The logit function will tend to +/- infinity for input values close
to 100% and 0% respectively. The logit or empirical logit function will be a
more appropriate choice than centred log-ratio (CLR) for non-OTU data.}
\usage{
normalise.logit(input)
}
\arguments{
\item{input}{Data frame of input compositional data to normalise. Input data
should be proportional 0-1.}
}
\value{
Data normalised using empirical logit. Proportions below 0.5 will be
negative, and output will tend to -/+ infinity for zero or 1 input.
}
\description{
Apply the standard logit normalisation to a data frame of omics input data.
This is intended to convert compositional data, e.g. proportional data in the
range 0..1, to Euclidean space which is most appropriate for the linear
models. The logit function will tend to +/- infinity for input values close
to 100% and 0% respectively. The logit or empirical logit function will be a
more appropriate choice than centred log-ratio (CLR) for non-OTU data.
}
\examples{
\dontrun{
normalise.logit(data.proportional)
}
}
