% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_analysis.R
\name{permanova.clusters}
\alias{permanova.clusters}
\title{Perform a permanova significance testing using the adonis function (from
vegan) upon consensus classes, as returned from the get.block.centroids
method. Limited to first two components.}
\usage{
permanova.clusters(consensus.summary)
}
\arguments{
\item{consensus.summary}{Consensus summary as generated by
get.block.centroids().}
}
\value{
Measurement of p-value for separation of clusters.
}
\description{
Perform a permanova significance testing using the adonis function (from
vegan) upon consensus classes, as returned from the get.block.centroids
method. Limited to first two components.
}
