% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrank.R
\name{plot.blockrank.scores}
\alias{plot.blockrank.scores}
\title{BlockRank Plotting Function}
\usage{
\method{plot}{blockrank.scores}(
  blockrank.i,
  nscores = 20,
  feature.font.size = 8,
  model = "",
  data_source = ""
)
}
\arguments{
\item{blockrank.i}{A list of numeric vectors containing BlockRank scores. One list item per data block}

\item{nscores}{The number of scores to display. Default 20.}

\item{feature.font.size}{Size of font of feature labels on y axis}

\item{model}{The type of model, used only in plot title.}

\item{data_source}{The name of the data source, used only in plot title.}
}
\value{
A ggplot2 plot object
}
\description{
Generates a bar plot of the top n blockrank scores, ordered highest to lowest,
for visualization purposes.
The default value 20 for nscores is not meaningful, just an example of approximately
how many features it is appropriate to visualise this way
}
\examples{
\dontrun{
plot.blockrank.scores(blockrank.score)
}
}
