% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{plot.feature.stability}
\alias{plot.feature.stability}
\title{Plot feature stability}
\usage{
\method{plot}{feature.stability}(selection.stability)
}
\arguments{
\item{selection.stability}{Stability data for features selected for a
component.}
}
\value{
Plotted stability data.
}
\description{
Plot feature stability for a summary overview.
}
\examples{
\dontrun{
plot.feature.stability(diablo.selection.stability(perf.diablo.analysis, comp=1, block=1))
}
}
