% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{train.permuted.model}
\alias{train.permuted.model}
\title{Perform a re-fit for a DIABLO model after permuting labels}
\usage{
train.permuted.model(data, design, data.labels, test.keepX)
}
\arguments{
\item{data}{Object containing input data blocks.}

\item{design}{DIABLO block relation design matrix.}

\item{data.labels}{Unpermuted data class labels.}

\item{test.keepX}{Array of values to test for sparse model training.}
}
\value{
Model performance balanced error rate.
}
\description{
Perform a full multi-omics model fitting and performance assessment after
permuting the labels associated with the classes. This can take some time.
This is useful when you want to assess whether a model is capable of
overfitting the data it was given.
}
