% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.calculate.correlation}
\alias{wgcna.calculate.correlation}
\title{Calculate the biweight midcorrelation between two matrices of data for the same set of samples.}
\usage{
wgcna.calculate.correlation(x, y)
}
\arguments{
\item{x}{Matrix x for the correlation.}

\item{y}{Matrix y for the correlation.}
}
\value{
A list containing the correlation values and the p values for each correlation between the two matrices.
}
\description{
Calculate the biweight midcorrelation between two matrices of data for the same set of samples.
}
