% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.module.colors}
\alias{wgcna.module.colors}
\title{Extract module colors from a constructed network.}
\usage{
wgcna.module.colors(network)
}
\arguments{
\item{network}{A constructed network to extract module colors from.}
}
\value{
A vector of color names equal in length and order to the genes in the
network.
}
\description{
Extract module colors from a constructed network.
}
