% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.plot.labelled.heatmap}
\alias{wgcna.plot.labelled.heatmap}
\title{Plot a labelled heatmap of correlations}
\usage{
wgcna.plot.labelled.heatmap(correlation)
}
\arguments{
\item{correlation}{A list containing the correlation values and p values for
the correlations.}
}
\description{
Plot a labelled heatmap of correlations, sorted by correlation value.
}
