% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.plot.module.membership.correlations}
\alias{wgcna.plot.module.membership.correlations}
\title{Generate all the plots of module membership correlation for all modules.}
\usage{
wgcna.plot.module.membership.correlations(
  module.colors,
  expr.module.correlation,
  expr.trait.correlation,
  include.modules = NULL,
  ignore.modules = c("grey")
)
}
\arguments{
\item{module.colors}{A list of module color names the same length as the
number of gene expression data in the correlations.}

\item{expr.module.correlation}{A correlation between the gene expression data
and the module eigengenes.}

\item{expr.trait.correlation}{A correlation between the gene expression data
and a trait from the sample metadata.}

\item{include.modules}{A list of module to include in the plots, or NULL for
all modules.}

\item{ignore.modules}{A list of module names to ignore when creating plots.}
}
\value{
A GO enrichment object which contains objects such as the
`enrichmentTable` which can be exported for later analysis of associated GO
terms for each module.
}
\description{
Generate all the plots of module membership correlation for all modules.
}
