% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.plot.network.dendro}
\alias{wgcna.plot.network.dendro}
\title{Plot a dendrogram for a constructed network with module colors plotted as a
horizontal stripe beneath.}
\usage{
wgcna.plot.network.dendro(network)
}
\arguments{
\item{network}{A constructed network to plot a dendrogram from.}
}
\description{
Plot a dendrogram for a constructed network with module colors plotted as a
horizontal stripe beneath.
}
