% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_mapping.R
\name{getGOGenes}
\alias{getGOGenes}
\title{Utility method to retrieve all gene annotated with a particular GO term, or
any of its children.}
\usage{
getGOGenes(go.id)
}
\arguments{
\item{go.id}{GO term.}
}
\value{
Symbol of gene annotated with this term.
}
\description{
Utility method to retrieve all gene annotated with a particular GO term, or
any of its children.
}
